+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestMORDataSource.scala
@@ -1243,6 +1243,57 @@ class TestMORDataSource extends HoodieClientTestBase with SparkDatasetMixin {
       roDf.where(col(recordKeyField) === 0).select(dataField).collect()(0).getLong(0))
   }
 
+  @ParameterizedTest
+  @CsvSource(Array(
+    "AVRO, AVRO, END_MAP",
+    "AVRO, SPARK, END_MAP",
+    "SPARK, AVRO, END_MAP",
+    "AVRO, AVRO, END_ARRAY",
+    "AVRO, SPARK, END_ARRAY",
+    "SPARK, AVRO, END_ARRAY"))
+  def testRecordTypeCompatibilityWithParquetLog(readType: HoodieRecordType,
+                                                writeType: HoodieRecordType,
+                                                transformMode: String): Unit = {
+    var (_, readOpts) = getWriterReaderOpts(readType)
+    var (writeOpts, _) = getWriterReaderOpts(writeType)
+    readOpts = readOpts ++ Map(HoodieStorageConfig.LOGFILE_DATA_BLOCK_FORMAT.key -> "parquet")
+    writeOpts = writeOpts ++ Map(HoodieStorageConfig.LOGFILE_DATA_BLOCK_FORMAT.key -> "parquet")
+    val records = dataGen.generateInserts("001", 10)
+
+    // End with array
+    val inputDF1 = transfrom(spark.read.json(
+      spark.sparkContext.parallelize(recordsToStrings(records).asScala, 2))
+      .withColumn("wk_tenant_id", lit("wk_tenant_id"))
+      .withColumn("ref_id", lit("wk_tenant_id")), transformMode)
+    inputDF1.write.format("org.apache.hudi")
+      .options(writeOpts)
+      .option(DataSourceWriteOptions.OPERATION_OPT_KEY, DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
+      .option(DataSourceWriteOptions.TABLE_TYPE_OPT_KEY, DataSourceWriteOptions.MOR_TABLE_TYPE_OPT_VAL)
+      // CanIndexLogFile=true
+      .option(HoodieIndexConfig.INDEX_TYPE_PROP, IndexType.INMEMORY.name())
+      .mode(SaveMode.Overwrite)
+      .save(basePath)
+    assertTrue(HoodieDataSourceHelpers.hasNewCommits(fs, basePath, "000"))
+
+    val snapshotDF1 = spark.read.format("org.apache.hudi")
+      .options(readOpts)
+      .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_SNAPSHOT_OPT_VAL)
+      .load(basePath + "/*/*/*/*")
+    snapshotDF1.show()
+    assertEquals(10, snapshotDF1.count())
+  }
+
+  def transfrom(sourceDF: DataFrame, transformed: String): DataFrame = {
+    transformed match {
+      case "END_MAP" => sourceDF
+        .withColumn("obj_ids", array(lit("wk_tenant_id")))
+        .withColumn("obj_maps", map(lit("wk_tenant_id"), col("obj_ids")))
+      case "END_ARRAY" => sourceDF
+        .withColumn("obj_maps", map(lit("wk_tenant_id"), lit("wk_tenant_id")))
+        .withColumn("obj_ids", array(col("obj_maps")))
+    }
+  }
+
   def getWriterReaderOpts(recordType: HoodieRecordType,
                           opt: Map[String, String] = commonOpts,
                           enableFileIndex: Boolean = DataSourceReadOptions.ENABLE_HOODIE_FILE_INDEX.defaultValue()):
